move wifiStatusStr helper
more wifiStatusStr
